# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NyayaSetu Legal Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List user documents
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Rename or recategorize document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # GUIDED LEGAL JOURNEYS
...
  /journeys/{id}:
    get:
      operationId: getJourney
      tags: [journeys]
      summary: Get journey details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Journey"
  /journeys/{id}/progress:
    get:
      operationId: getJourneyProgress
      tags: [journeys]
      summary: Get user progress for a journey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JourneyProgress"
    patch:
      operationId: updateJourneyProgress
      tags: [journeys]
      summary: Update journey progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JourneyProgressUpdate"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JourneyProgress"

  # GOVERNMENT SCHEMES
...
  /learning:
    get:
      operationId: listLearningContent
      tags: [learning]
      summary: List learning content
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Learning content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LearningContent"
  /learning/{id}:
    get:
      operationId: getLearningContent
      tags: [learning]
      summary: Get learning content detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningContent"
  /learning/{id}/bookmark:
    post:
      operationId: bookmarkContent
      tags: [learning]
      summary: Bookmark or unbookmark content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "200":
          description: Bookmark result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /legalaid:
    get:
      operationId: listLegalAid
      tags: [legalaid]
      summary: Search legal aid providers
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: area
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Legal aid providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalAidProvider"
  /legalaid/{id}:
    get:
      operationId: getLegalAidProvider
      tags: [legalaid]
      summary: Get legal aid provider details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalAidProvider"

  # AI CHAT
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Document:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        originalName:
          type: string
        category:
          type: ["string", "null"]
        mimeType:
          type: string
        size:
          type: integer
        filePath:
          type: string
        createdAt:
          type: string
      required: [id, name, originalName, mimeType, size, filePath, createdAt]
    DocumentUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
    Journey:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        estimatedDays:
          type: integer
        eligibility:
          type: string
        requiredDocuments:
          type: array
          items:
            type: string
        applicableLaws:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/JourneyStep"
      required: [id, title, description, category, estimatedDays, eligibility, requiredDocuments, applicableLaws, steps]
    JourneyStep:
      type: object
      properties:
        stepNumber:
          type: integer
        title:
          type: string
        description:
          type: string
        tips:
          type: ["string", "null"]
      required: [stepNumber, title, description]
    JourneyProgress:
      type: object
      properties:
        journeyId:
          type: integer
        currentStep:
          type: integer
        completedSteps:
          type: array
          items:
            type: integer
        completionPercent:
          type: integer
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
      required: [journeyId, currentStep, completedSteps, completionPercent]
    JourneyProgressUpdate:
      type: object
      properties:
        currentStep:
          type: integer
        completedSteps:
          type: array
          items:
            type: integer
      required: [currentStep, completedSteps]
...
    LearningContent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        type:
          type: string
        tags:
          type: array
          items:
            type: string
        isBookmarked:
          type: boolean
        createdAt:
          type: string
      required: [id, title, summary, content, type, tags, isBookmarked, createdAt]
    BookmarkInput:
      type: object
      properties:
        bookmarked:
          type: boolean
      required: [bookmarked]
    LegalAidProvider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: ["string", "null"]
        address:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        practiceAreas:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        description:
          type: string
      required: [id, name, type, city, state, address, practiceAreas, languages, description]
...