# lib/api-zod/src/generated/types/journeyProgress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NyayaSetu Legal Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface JourneyProgress {
  journeyId: number;
  currentStep: number;
  completedSteps: number[];
  completionPercent: number;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}